项目结构调整
1.
将每位组员写的view函数和views.py分开，避免引入不必要的conflicts（比如xyh之前push的放置到change_persona
l_info.py中）；
2. 将系统组件的诸import写到head.py中，之后写view只要import head即可获得全部基础组件（若有缺失请往里添加）；
3.
将url配置分散到每个app，在顶层的djcode/djcode/urls.py中include了我们IMS子系统的djcode/IMS/url
s.py，所有IMS下辖的url pattern都通过localhost:8000/ims/...来访问。
