+++ djcode/IMS/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
 # Register your models here.
-from models import Student_user
+from IMS.models import Student_user
 
 admin.site.register(Student_user)
