+++ djcode/IMS/templates/login.html
@@ -12,7 +12,7 @@
 <div class="container">
     <div class="row" align="middle">
         <div class="col" align="middle">
-            <form role="form" action="../user_auth/" method="post" id="login_form"> {% csrf_token %}
+            <form role="form" action="user_auth/" method="post" id="login_form"> {% csrf_token %}
                 <div class="input-prepend">
                     <span class="add-on"><span class="icon-user"></span></span>
                     <input class="span2" placeholder="请输入ID" type="text" name="userid" />
