+++ djcode/IMS/urls.py
@@ -16,8 +16,16 @@ Including another URLconf
     2. Add a URL to urlpatterns:  url(r'^blog/', include(blog_urls))
 """
 from django.conf.urls import include, url
-from IMS import views
+from IMS import views, change_personal_info
 
 urlpatterns = [
-    url(r'^hello/', views.hello),
+    url(r'^$', views.loggingin),
+    #url(r'^hello/', views.startup),
+    url(r'^add_user/', views.add_user),
+    url(r'^user_added/', views.user_added),
+    url(r'^user_auth/', views.user_auth),
+    url(r'^home/', views.home),
+    url(r'^logout/', views.loggingout),
+    url(r'^changeStudentInfo/', change_personal_info.changeStudentInfo), #by xyh
+    url(r'^changePasswd/', change_personal_info.changePasswd), #by xyh
 ]
